pyproject.toml
[tool.poetry]
name = "mirror"
version = "0.1.0"
description = "Mirror — AI Companion Bot"
authors = ["Mirror Team"]
packages = [{ include = "mirror" }]

[tool.poetry.dependencies]
python = "^3.12"

# Core
fastapi = "^0.115"
uvicorn = { extras = ["standard"], version = "^0.34" }
pydantic = "^2.10"
pydantic-settings = "^2.7"

# Database
sqlalchemy = { extras = ["asyncio"], version = "^2.0" }
asyncpg = "^0.30"
alembic = "^1.14"
redis = { extras = ["asyncio"], version = "^5.2" }

# Vector DB
qdrant-client = "^1.13"

# Telegram
aiogram = "^3.17"

# LLM / AI
openai = "^1.60"
anthropic = "^0.43"
langgraph = "^0.2"
haystack-ai = "^2.9"

# Astrology
kerykeion = "^4.20"

# Task Queue
celery = { extras = ["rabbitmq"], version = "^5.4" }

# Utils
geopy = "^2.4"
pytz = "^2024.1"
httpx = "^0.28"
pypdf = "^5.1"
python-docx = "^1.1"
beautifulsoup4 = "^4.12"
lxml = "^5.3"
python-multipart = "^0.0.20"

# Messaging
nats-py = "^2.9"

# Auth
PyJWT = "^2.9"

# Logging / Monitoring
structlog = "^24.4"
sentry-sdk = { extras = ["fastapi"], version = "^2.19" }
prometheus-fastapi-instrumentator = "^7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-mock = "^3.14"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
